+++ rust/ommx/src/mps/tests.rs
@@ -3,4 +3,4 @@
 mod edge_cases;
 mod invalid_cases;
 mod read_tests;
-mod roundtrip_tests;
+mod write_arbitrary;
