+++ rust/ommx/src/mps/tests/write_arbitrary.rs
